refactoring, rest endpoints changed, students.js file split -> form & list, code clean up
