+++ src/main/java/pl/arturzaczek/demoSchool/controllers/rest/StudentRestController.java
@@ -1,11 +1,9 @@
 package pl.arturzaczek.demoSchool.controllers.rest;
 
-import org.omg.CORBA.portable.Delegate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import pl.arturzaczek.demoSchool.model.entities.Grade;
 import pl.arturzaczek.demoSchool.model.entities.Student;
 import pl.arturzaczek.demoSchool.model.repositories.StudentRepository;
 
@@ -13,9 +11,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-/**    /students TODO zmienić mapowanie kontrolerów na /rest/ a pozniej Dodac @RequestMapping nad klase
- */
 @RestController
+@RequestMapping("/rest")
 public class StudentRestController {
     StudentRepository studentRepository;
     Logger logger = LoggerFactory.getLogger(StudentRestController.class);
@@ -25,22 +22,7 @@ public class StudentRestController {
         this.studentRepository = studentRepository;
     }
 
-/**    /students TODO zmiana endpointow zgodnie z konwencja
- *
-A resource can be a singleton or a collection. For example, “customers” is a
-    collection resource and “customer” is a singleton resource (in a banking domain).
-    We can identify “customers” collection resource using the URI “/customers”.
-    We can identify a single “customer” resource using the URI “/customers/{customerId}”.
-
-    A resource may contain sub-collection resources also. For example,
-    sub-collection resource “accounts” of a particular “customer” can be identified using the
-    URN “/customers/{customerId}/accounts” (in a banking domain). Similarly, a singleton resource
-    “account” inside the sub-collection resource “accounts” can be identified
-    as follows: “/customers/{customerId}/accounts/{accountId}”.
-*/
-
-
-    @GetMapping("/rest/students")
+    @GetMapping("/students")
     public List<Student> getStudents(){
         logger.debug("getStudents()");
         List<Student> studentList = studentRepository.findAll();
@@ -48,14 +30,14 @@ A resource can be a singleton or a collection. For example, “customers” is a
         return studentList;
     }
 
-    @PostMapping("/addStudent")
+    @PostMapping("/student")
     public void addStudent(@RequestBody Student student){
         System.out.println("/addStudent " + student);
         logger.debug("addStudent() rest", student.toString());
         studentRepository.save(student);
     }
 
-    @GetMapping("/rest/add5Students")
+    @GetMapping("/add5Students")
     public void add5Students(){
         List studentlist = new ArrayList<Student>();
         studentlist.add(new Student("Artur", "Aaaaa"));
@@ -66,23 +48,14 @@ A resource can be a singleton or a collection. For example, “customers” is a
         logger.debug("add5Students() rest",studentlist );
         studentRepository.saveAll(studentlist);
     }
-    @GetMapping("/add1StudentWithGrade")
-    public void add1StudentWithGrade(){
-        Student student = new Student("Artur", "Z ocenami");
-        System.out.println("student " + student);
-        logger.debug("add1StudentWithGrade() rest " + student.toString());
-        student.getGradeList().add(new Grade(80));
-        System.out.println("student " + student);
-        studentRepository.save(student);
-    }
-    @DeleteMapping("/rest/student/{student_id}")
+    @DeleteMapping("/student/{student_id}")
     public String deleteStudentById(@PathVariable String student_id){
         System.out.println("student_id " + student_id);
-        logger.debug("deleteStudentById() rest " + student_id.toString());
+        logger.debug("deleteStudentById() rest " + student_id);
         Long long_id = Long.parseLong(student_id+"");
         Optional<Student> byId = studentRepository.findById(long_id);
         if(!byId.isPresent()){
-            logger.warn("deleteStudentById() rest " + student_id.toString());
+            logger.warn("deleteStudentById() rest " + student_id);
             return "not found: " + student_id;
         }
         studentRepository.deleteById(long_id);
