+++ src/main/resources/templates/students/addStudent.html
@@ -7,7 +7,6 @@
     <th:block th:include="fragments/imports :: head"></th:block>
     <th:block th:include="fragments/imports :: student-liks"></th:block>
 </head>
-
 <body>
 <th:block th:include="fragments/imports :: nav"></th:block>
 <div class="main-wrapper">
@@ -44,7 +43,7 @@
             </div>
             <label for="birth-date" class="form-label">Data urodzenia</label>
             <div class="input-group mb-3">
-                <input type="date" id="birth-date" class="form-control datepicker rounded-right" placeholder="data urodzenia"  min="1950-01-01" max="2005-12-31">
+                <input type="date" id="birth-date" class="form-control datepicker rounded-right" placeholder="data urodzenia"  min="1960-01-01" max="2005-12-31">
                 <span class="input-group-text">Data urodzenia</span>
                 <div class="valid-feedback">
                     Ok!
